+++ components/ui/organisms/index.ts
@@ -1,0 +1,1 @@
+export { default as Article } from './Article';
